ugc_service/requirements.txt
aiohttp==3.8.6
aiokafka==0.8.1
fastapi==0.104.0
gunicorn==21.2.0
kafka-python==2.0.2
motor==3.3.2
orjson==3.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
pymongo==4.5.0
python-jose==3.3.0
sentry-sdk==1.34.0
uvicorn==0.23.2
